Remove 'cluster' and 'role' from node adopt
We are abstaining from permitting a node to adopted directly into a
cluster. This is gonna be dangerous because adding new members to a
cluster means a lot of things rather than setting the cluster and index
property. It will affect placement, load-balancing and many other
policies. Before we have a plan to do them, we limit this api to adopt a
node as orphan node only.

Change-Id: I6e58932bb350906f3b04bc7e124ead0c0b181c11