Set version=auto when creating docker client
When creating a docker client using docker library, a specific
docker api version (i.e. 1.26.0 in python-docker-2.4.2) is used
by default. However the version of docker included in distros
can be different. Using version='auto', the client detects the
api version in the server and uses it.

This change sets version to 'auto' so that is automatically
detected as kolla [1] and others do.

[1] https://github.com/openstack/kolla/blob/master/kolla/image/build.py#L135

Change-Id: I42d2b4291ca8a8a132b9b7b72002f7ddd714fe5f