revising node recovery logic
this patch revises node recovery logic which dumped data to action.outputs to help us
tell whether is had been recreated. Now we will dump it do node.data cause the logic is in engine node.

And for the nodes was in WARNING status, we will mark it as ACTIVE either if its physical status is fine,
and all these nodes will be handled in lb policy if needed.

Change-Id: I56234cb8d8e7d54bcd9a1f97f6d486318a332286