2nd patch to reverse GC process
We will try to delete the dependents first and then mark the actions'
status as Failed since the owner of the actions went down.

After that, we will try to release node lock and cluster lock(cluster scope
or node scope)

One problem is that if the dead engine hold a cluster action with a number
of sub-actions belong to mult engines, we will mark them all as 'FAILED'.

The starting point for this design is to mark all sub-actions which status
are 'READY' to 'FAILED' so that the scheduler will not try to acquire them and
process them.

Change-Id: I53a5e1d19f4adb04580627cb6306c43efe610553