fix the bug that policy validate cannot work when creatting a policy
It can be created successfully when I create a policy with an
invalid subnet. And I find the default value is False of the parameter
validate_props,so that the policy validate() cannot validate deeply.

To fix the bug, I change the validate_props from False to True in the
code of policy_create().

Change-Id: I727b3a0d74ff69d8cd797b8a5af8d6f40a8f5fb9
Closes-Bug: 1716812