move physical_id check from engine object to profile layer
this patch revise node deletion process to move physical_id check
to profile layer, so that users could have more change to customize
this process.

for some cases we may not have the physical_id, but the server is
running in nova, or somewhere. In this cases, we may need to delete
server by server_name or metadata.

Change-Id: If0759561cca32aa39efa3245960d03a70219b963