Name generator util function
This adds a name generator function for cluster nodes. When provided
with a fmt that has "$nR" or "$nI" in it, the util function will perform
in-place substitution of the format string. The "$R" will be replaced
with a random string, the "$I" will be replaced with the node index.
The "n" in the format string is optional. When specified, it creates the
random string of given length, or a string representation of node index
left padded with 0s.

Change-Id: I6dfd95be4a315917bbc4920678921618df4792c0