Fix the bug that cannot create a cluster
It report the error with "Field `user' cannot be None" when
I create a cluster.
I found the RequestContext() code had changed from 'user' to
'user_id' in the commit 66e8486c27b30e73d60899537fcf6c5e97ea261a,
but in action/base.py, the params were not to changed. So cause
the 'user' is None and cannot create a cluster successfully.
I modify the 'user' to 'user_id' to match the RequestContext()
of context.py.

Closes-Bug: #1740377
Change-Id: Id0e12c31b4e107bb0c69c394e81d3e0ef10462f9