Fix the bug that could not create a profile with Chinese availability_zone
There has an error message with " 'ascii' codec can't encode characters in
 position 0-1: ordinal not in range(128)" when I create a profile with
Chinese availability_zone.
Then I found the str(value) send an error if the value is u'\u53ef\u7528\u52ff\u5220'
in schema.py when validate the profile. I add a if to judge the value is
unicode or not.

Closes-Bug: #1739877
Change-Id: I2a6568729a6280cc29c8f58f67a290b289714aac