Fix the bug that couldn't create a receiver
When I create a receiver for a cluster,there has
failed with the error message "HttpException:
Internal Server Error"

Then I found there has not the function get_schema()
in ClusterActionName class and ReceiverType class,
so that raise NotImplementedError() from the class
FieldType of oslo_versionedobjects/fields.py.

Then I add the get_schema() in ClusterActionName and
ReceiverType and fix this bug.

Closes-Bug: #1744426
Change-Id: Ie29e1dffdedb37b117ec698fba434084f5d5f831