Update sdk connection, tests and isoformat
We can simplify connection creation now. The Connection constructor
understands not to load yaml or env vars if it doesnt' receive a cloud
argument - so this can all be done in one step.

Update the mocks in the tests to work for the new calling pattern.

Finally - isoformat wasn't processing UTC+00:00. There are some
dark issues with how isoformats get passed around OpenStack, but this
fixes the unittest matching to account for UTC+00:00.
openstacksdk has been working on merging the code from shade and
os-client-config. Part of this will result in the removal of the
Profile object in favor of the CloudRegion object from the new
openstack.config.

This updates the Connection construction code to code that should work
with both old and new versions of openstacksdk.

Temporarily mask some jobs to get this in. Will need to revise the
tempest plugin to fix gate jobs later.

Related-Bug: #1681620
Co-Authored-By: tengqm <tengqim@cn.ibm.com>
Change-Id: I10c74ca48b1ddb848a5a68cc4360431a21e0a2cc