Catch exceptions when updating service
service_manage_report is periodically called by the engine service
timer.  If an exception occurs during the execution of
service_manage_report, the timer is killed.  This leaves the engine
service to believe it is dead.

With this change any exceptions encountered during service update will
be caught and logged without killing the timer.

The same exception handling is added to service_manage_cleanup.

Closes-Bug: #1763490

Change-Id: Id1ba7f9c815cfebe2b2d269640a6544ababd7ae8