Do not for force-reinstall when upgrading packages
Using fore-reinstall with upgrade will cause all
dependencies to always be reinstalled. This can
cause issues when other components install Senlin
as a dependency.

In this case we are causing issues because we are
re-installing certify, a file that is modified by
devstack after installation to make sure that the
system has the appropriate ssl certificates.

See the function fix_system_ca_bundle_path in devstack
for more information.

Change-Id: Ibb6578f45f9277f97ca61b60714c8ce2b4eacb61
Closes-Bug: #1769541