Fix broken SDK exception test
In openstacksdk 0.14.0 release, a bug related to SDK exception was
fixed: https://review.openstack.org/#/c/571101/. With that change a SDK
exception will contain the detailed message only if the message string
is equal to 'Error'.

This commit fixes the test_parse_exception_http_exception_no_details to
use 'Error' as the exception message to make the test case pass.

Change-Id: I61368430e0f8e8316e13276452da05041920ce4e