Add cluster option cluster.stop_node_before_delete
This adds 'cluster.stop_node_before_delete' as a new boolean cluster
config option. If true, a cluster will issue a stop to a node and then a
delete for all cluster operations that trigger a node delete. If not
set, the config option defaults to False and the cluster will perform a
regular node delete. The cluster operations affected by this change are
cluster delete, scale-in and resize.

Change-Id: I2248419cc7b1e359c9a189fae97d7883bab3ec84