Fix stop node before delete error handling
Log warning message instead of returning error when a cluster is trying
to stop a node but fails to do so.  Then proceed to delete the node.
This change is needed because in case of a failure to stop a node, the
node should still be deleted to clean it up.

Change-Id: I345ddb47d4f7715bfc7251c4eb2a520bbbcabdf3