Update scaling policy logic to be applied before action acceptance
This patch alters the cluster_scale_in and cluster_scale_out actions to
no longer place the action into the actions table when a conflict is
detected. This behavior is an improvement on the old way actions are
processed as the requester will now receive immediate feedback from the API
when an action cannot be processed. This change  also honors the scaling
action cooldown in the same manner by erring via the API when a scaling
action cannot be processed due to cooldown.

Depends-On: https://review.openstack.org/#/c/602460/
Implements: blueprint scaling-action-acceptance
Change-Id: If0dcf5e427d3d6973d2c5e52fada8a6c925240d5