Support multiple detection types in health policy
This patchset changes the existing health policy to allow
NODE_STATUS_POLLING to be combined with NODE_STATUS_POLL_URL to provide
multiple detection types.

The existing implementation of NODE_STATUS_POLLING was changed to
directly perform a do_check on the node instead of doing a cluster check
operation. This was done to avoid locking the cluster as part of the
health check. When the cluster check was performed during a health
check, there was a potential lock contention problem if the user was
already performing other operations on the same cluster.

Implements: blueprint multiple-detection-modes
Change-Id: Iba0b043cc582e668a9066b0586ca8a0b201040e4