Fix test_run_workflow unittests
The test_run_workflow unittests occasionally fail because one the call
args is a string representation of a dictionary and the order of its
keys could be different than that of the string representation of the
expected dictionary.

This ordering problem is fixed by turning the actual call argument into
a json object then dumping it as a string sorted by key.  The expected
dictionary is also dumped into a string sorted by key.  Then both sorted
values are compared.

Change-Id: I4b3cb5f713694f8667050fc7d5d54cfb597d216a