Propagate node creation errors
When a node fails to create due to an exception returned by
openstacksdk, the exception message is not available to the user after
the action is marked as failed.  The action status_reason is set to the
exception message inside engine/node.py do_create.  However it is
overwritten again in engine/actions/base.py ActionProc.

The fix is to return the error message out of engine/node.py do_create
and so that engine/actions/base.py ActionProc can set the status_reason
to the error message.

Change-Id: Id34d40be3752b3b4f7e12eb28b6fa88741b13d78