Add force to action_update
This patch adds the force query parameter to action_update.
When force is set to true on an action cancel request the
action status will be updated in the database to cancelled
rather than signaled to cancel. This will also update any
child actions to update their status in the database to
cancelled.

Change-Id: I45ab4a6ad57e3bfe791da1d2e70d3b2ed3ad9a21