Fix member address selection in lbaas driver
Currently when Senlin adds a member to load balancer's pool, it just uses
the first address of the node. In a dual stack network, if the first
address does not match with the subnet ip version, the member will become
unaccessible from the load balancer. This patch ensures the right address
will be picked according to the subnet ip version.

Change-Id: I0322184dfdb46a7560c20e85452d84d2ffb83a1c
Closes-Bug: #1813089