Only update necessary metadata
When issuing a metadata update, it is better to simply update the
key/value pairs that we need to update rather than completely deleting
and recreating the dictionary from scratch. It avoids problems such as
when parts of the server metadata have been tagged as immutable by
admins. Metadata keys that cannot be updated due to permissions are
simply skipped over now.

Change-Id: I6f928ce992d48e062ffd2017a665dc16641ac4b3