Add webhook v2 support
Previously webhook API introduced microversion 1.10 to allow callers to
pass arbritary data in the body along with the webhook call.  This was
done so that webhooks would work with aodh again.  However, aodh and
most webhook callers cannot pass in the header necessary to specify the
microversion.  Thus, I'm introducing webhook v2 here so that webhooks
behave like in microversion 1.10 but without the need to specify that
microversion header.

Closes-Bug: #1828856
Change-Id: I7f5cdb2ea19c6ca1a9ea68e64515804f219a73ee