Add cluster_id to action filter in API
Current, an action was created haven't information of a cluster.
When a specific action is created, we do not know what action this action takes
in a cluster and when we execute the query API:

GET http://senlin_host:8777/actions

so many actions are returned even though we can limit using parameters
such as name, target, action, status combined with limit and marker.
But, we can't just get the actions executed in a particular cluster.
So, I want us to be able to do this by enabling the query with a cluster_id parameter.
To accomplish this, I need to modify the schema of
action models and action tables in the database of senlin service.

Change-Id: I49d172414eecf9158192cf0869abb91822ee2226