Fix leaking resources during cluster recover
If a node recover was in progress when a cluster delete was
called it would cause resources to be created, but unable
to update the Node object as it no longer exists.

With this patch if Node.update fails because the object
no longer exists. Delete all resources created during
the failed operation.

Change-Id: I95a8b5791eee5d334f8841bebe52f37f6f5eb2e5