Fix requirements, update hacking
hacking already pins flake8, so do not add it, especially in conflict
with the hacking version. hacking also pins pycodestyle, so remove that
as well.

The repo is Python 3 now, so update hacking for a python3 version.

Blacklist new failures or fix them.

Convert own hacking checks to new flake8 API, update lower-constraints
for this.

Change-Id: I2928b9d764f61a53a549a9eb850c7809bdbb6f74