Fix devstack problems causing failed CI tests
- Use uwsgi binary from path:
Devstack projects have been fixed to run uwsgi binary from the path
instead of the project specific directory [1].  This applies the same
fix to senlin-api so that it makes uwsgi invocation to find it in the
path.

- Set keystone auth interface to public:
The default identity interface was changed from admin to internal in
[2].  Since devstack only creates admin and public endpoints for
identity service, senlin fails to talk to identity service without the
explicit setting to use public interface during authentication.

[1] https://review.opendev.org/#/c/577779/
[2] https://review.opendev.org/#/c/662734/

Change-Id: Ia8f56a7016eace5b7ce70dc661729e19284c3733