Find security group profiles by project scope
If a security group exists with the same name in multiple projects, a
profile will not be able to created using that security group name.
This change will try to scope the security group to the user's project
during validation.  If the security group can not be found using the
project scope, it will try to find it without the project scope.

Change-Id: I595b3c348ce8133a273d8e26c664fffc32694705