Fixed service manage not cleaning properly
When spliting the engine up into multiple services
the cleanup job was left on the conductor, but in
reality each service needs to handle their own clean
up, as the service startup order is never guaranteed.

This patch refactors the service workflow and moves
all cleanup logic to the base class.

Closes-Bug: 1975440
Change-Id: I718910a124780804ec8e14c2e99697025e4c4df0