fix: SCALE_IN action not honoring specified node count
The SCALE_IN action in the lb_policy module was removing only
one node from the load balancer, regardless of the specified
count when the count was greater than 1. This was due to a
hard-coded count of 1 in the _get_delete_candidates function.

The root cause has been addressed by updating the
_get_delete_candidates function to dynamically handle the count
for SCALE_IN. The count is now retrieved from the action inputs,
and if not specified or invalid, it defaults to 1.

Closes-Bug: #2048099
Change-Id: I9ff12dec26f4c558ef5ef84de87668e9668cec80